Rename `exportConfig` as `config`
